Moved Packet file to ECP directory and fixed all wrong paths
