+++ src/app/data/IHandler.ts
@@ -1,4 +1,4 @@
 import { TcpServer } from "../../server/types";
-import Packet from "../utils/parser/Packet";
+import Packet from "../encryptedChatProtocol/Packet";
 
 export type HandlerCb<T extends Packet> = (req: T, res: TcpServer.IResponse) => void;
