+++ src/app/encryptedChatProtocol/requestPackets/Login.ts
@@ -1,6 +1,6 @@
 import { IBuilder } from "../../common/IBuilder";
 import { AuthAttributs, PacketType } from "../commonTypes";
-import Packet from "../../utils/parser/Packet";
+import Packet from "../Packet";
 
 export default class LoginRequestPacket extends Packet {
     readonly userAttributs: AuthAttributs;
