+++ src/app/encryptedChatProtocol/requestPackets/RequsetPacket.ts
@@ -1,5 +1,5 @@
 import { PacketType } from "../commonTypes";
-import Packet from "../../utils/parser/Packet";
+import Packet from "../Packet";
 
 export default abstract class RequestPacket extends Packet {
     constructor(type: PacketType, packetId?: string) {
