--- src/app/utils/parser/IHandler.ts
@@ -1,3 +1,0 @@
-import Packet from "./Packet";
-
-export type Handler<T> = (req: T, res: unknown/*IResponse*/) => void;
